+++ scanpipe/tests/test_pipelines.py
@@ -702,7 +702,9 @@ class PipelinesIntegrationTest(TestCase):
         pipeline_name = "inspect_manifest"
         project1 = Project.objects.create(name="Analysis")
 
-        input_location = self.data_location / "Django-4.0.8-py3-none-any.whl.ABOUT"
+        input_location = (
+            self.data_location / "manifests" / "Django-4.0.8-py3-none-any.whl.ABOUT"
+        )
         project1.copy_input_from(input_location)
 
         run = project1.add_pipeline(pipeline_name)
@@ -722,7 +724,7 @@ class PipelinesIntegrationTest(TestCase):
         pipeline_name = "inspect_manifest"
         project1 = Project.objects.create(name="Analysis")
 
-        input_location = self.data_location / "toml.spdx.json"
+        input_location = self.data_location / "manifests" / "toml.spdx.json"
         project1.copy_input_from(input_location)
 
         run = project1.add_pipeline(pipeline_name)